# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Painless GrowthOS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/capture-endpoints:
    get:
      operationId: listCaptureEndpoints
      tags: [capture]
      summary: List inbound lead-capture endpoints
      responses:
        "200":
          description: Capture endpoints, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaptureEndpoint"
    post:
      operationId: createCaptureEndpoint
      tags: [capture]
      summary: Create an inbound lead-capture endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaptureEndpointInput"
      responses:
        "201":
          description: Created endpoint (token included)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptureEndpoint"
        "400":
          description: Invalid endpoint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/capture-endpoints/{id}:
    patch:
      operationId: updateCaptureEndpoint
      tags: [capture]
      summary: Update name, mapping, source, or active flag
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaptureEndpointUpdate"
      responses:
        "200":
          description: Updated endpoint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptureEndpoint"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteCaptureEndpoint
      tags: [capture]
      summary: Deactivate a capture endpoint
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "204":
          description: Deactivated
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/capture-endpoints/preview:
    post:
      operationId: previewCaptureMapping
      tags: [capture]
      summary: Apply a field mapping to a test payload without saving anything
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaptureMappingPreviewRequest"
      responses:
        "200":
          description: Mapped lead fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptureMappingPreview"
        "400":
          description: Invalid mapping
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  parameters:
    IdPathParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    CaptureFieldMapping:
      type: object
      description: External payload field name -> MogulForge lead field name
      additionalProperties:
        type: string
    CaptureEndpointInput:
      type: object
      required: [name, mapping]
      properties:
        name:
          type: string
        mapping:
          $ref: "#/components/schemas/CaptureFieldMapping"
        defaultSource:
          type: string
    CaptureEndpointUpdate:
      type: object
      properties:
        name:
          type: string
        mapping:
          $ref: "#/components/schemas/CaptureFieldMapping"
        defaultSource:
          type: string
        isActive:
          type: boolean
    CaptureEndpoint:
      type: object
      required: [id, name, token, mapping, defaultSource, isActive, receivedCount, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        token:
          type: string
        mapping:
          $ref: "#/components/schemas/CaptureFieldMapping"
        defaultSource:
          type: string
        isActive:
          type: boolean
        lastReceivedAt:
          type: string
          nullable: true
        receivedCount:
          type: number
        createdAt:
          type: string
        url:
          type: string
          description: Full inbound POST URL for this endpoint
        embedSnippet:
          type: string
          description: Paste-in script tag for the site form listener
    CaptureMappingPreviewRequest:
      type: object
      required: [mapping, payload]
      properties:
        mapping:
          $ref: "#/components/schemas/CaptureFieldMapping"
        payload:
          type: object
          additionalProperties: true
    CaptureMappingPreview:
      type: object
      required: [unmapped]
      properties:
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        addressLine1:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        postalCode:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
        source:
          type: string
          nullable: true
        campaign:
          type: string
          nullable: true
        externalId:
          type: string
          nullable: true
        unmapped:
          type: array
          items:
            type: string
...