# lib/api-zod/src/generated/types/captureEndpoint.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Painless GrowthOS API
 * OpenAPI spec version: 0.1.0
 */
import type { CaptureFieldMapping } from './captureFieldMapping';

export interface CaptureEndpoint {
  id: string;
  name: string;
  token: string;
  mapping: CaptureFieldMapping;
  defaultSource: string;
  isActive: boolean;
  lastReceivedAt?: string | null;
  receivedCount: number;
  createdAt: string;
  /** Full inbound POST URL for this endpoint */
  url?: string;
  /** Paste-in script tag for the site form listener */
  embedSnippet?: string;
}