# lib/api-zod/src/generated/types/captureEndpointInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Painless GrowthOS API
 * OpenAPI spec version: 0.1.0
 */
import type { CaptureFieldMapping } from './captureFieldMapping';

export interface CaptureEndpointInput {
  name: string;
  mapping: CaptureFieldMapping;
  defaultSource?: string;
}